Move mostly client-side modules to 'client' package
