+++ pylambder/__init__.py
@@ -1,6 +1,6 @@
 """Main pylambder package"""
 
-import pylambder.base as base
+import pylambder.client.base as base
 
 APP_NAME = "pylambder"
 __version__ = "0.2.1"
